Update gitQuizMain.c
Changes made:
- Deleted comment on line 1:
The comment was deleted on Line 1 because the instruction was already followed through and completed.
- Added else condition:
As instructed by the comment on line 1, an else condition was added on line 1 which includes any other input that is not 'y'. When an input is given that is not 'y', the program outputs the message "You do not know how to use Git!".